Extract BootTidal class
