+++ lib/repl.js
@@ -2,6 +2,7 @@
 
 var fs = require('fs');
 const Ghci = require('./ghci')
+const BootTidal = require('./boot-tidal')
 var defaultBootFileName = 'BootTidal.hs';
 var CONST_LINE = 'line'
 var CONST_MULTI_LINE = 'multi_line'
@@ -102,21 +103,8 @@ export default class REPL {
 
     }
 
-    getBootTidalPath(rootDirectories) {
-        const configuredBootFilePath = atom.config.get('tidalcycles.bootTidalPath');
-        if (configuredBootFilePath) return configuredBootFilePath
-
-        const currentDirectoryPath = rootDirectories.length > 0 ?
-            rootDirectories[0].path + '/' + defaultBootFileName :
-            null;
-
-        if (fs.existsSync(currentDirectoryPath)) return currentDirectoryPath;
-
-        return  __dirname + '/' + defaultBootFileName;
-    }
-
     initTidal() {
-        const bootPath = this.getBootTidalPath(atom.project.rootDirectories);
+        const bootPath = BootTidal.getPath()
         const blocks = fs.readFileSync(bootPath)
             .toString()
             .split('\n\n')
