+++ spec/repl-spec.js
@@ -1,4 +1,5 @@
 const REPL = require('../lib/repl.js')
+const BootTidal = require('../lib/boot-tidal')
 const fs = require('fs')
 
 describe('repl', () => {
@@ -11,7 +12,7 @@ describe('repl', () => {
   describe('init tidal', () => {
 
     it('should load code blocks on boot', () => {
-      spyOn(repl, 'getBootTidalPath').andReturn("lib/BootTidal.hs")
+      spyOn(BootTidal, 'getPath').andReturn("lib/BootTidal.hs")
       spyOn(repl, 'tidalSendLine')
       spyOn(repl, 'tidalSendExpression')
 
@@ -24,7 +25,7 @@ describe('repl', () => {
     })
 
     it('should send :set lines as single lines on boot', () => {
-      spyOn(repl, 'getBootTidalPath').andReturn("lib/BootTidal.hs")
+      spyOn(BootTidal, 'getPath').andReturn("lib/BootTidal.hs")
       spyOn(repl, 'tidalSendLine')
       spyOn(repl, 'tidalSendExpression')
 
@@ -34,29 +35,6 @@ describe('repl', () => {
       expect(repl.tidalSendLine.calls[1].args[0]).toBe(':set prompt ""')
       expect(repl.tidalSendLine.calls[2].args[0]).toBe(':set prompt-cont ""')
       expect(repl.tidalSendLine.calls[3].args[0]).toBe(':set prompt "tidal> "')
-    })
-  })
-
-  describe('boot file sequence', () => {
-    let rootDirectories = [{ path: '/current/directory' }]
-
-    it(`should choose current directory BootTidal.hs if it exists`, () => {
-      spyOn(fs, 'existsSync').andReturn(true)
-
-      expect(repl.getBootTidalPath(rootDirectories)).toBe('/current/directory/BootTidal.hs')
-    })
-
-    it(`should choose custom boot file when provided`, () => {
-      atom.config.set('tidalcycles.bootTidalPath', '/custom/directory/BootTidal.hs')
-      spyOn(fs, 'existsSync').andReturn(true)
-
-      expect(repl.getBootTidalPath(rootDirectories)).toBe('/custom/directory/BootTidal.hs')
-    })
-
-    it(`should choose default boot file when no custom provided and no file in current directory`, () => {
-      spyOn(fs, 'existsSync').andReturn(false)
-
-      expect(repl.getBootTidalPath([])).toContain('/lib/BootTidal.hs')
     })
   })
 })
